fix: move TDB reciprocity test from kani_verif.rs to tests/epoch.rs
The #[test] formal_epoch_reciprocity_tdb was inside src/epoch/kani_verif.rs
which is gated behind #[cfg(kani)]. It never ran during normal cargo test.

Moved to tests/epoch.rs using public API (to_parts() instead of direct
field access). Removed the dead test and its comments from kani_verif.rs,
along with the now-unused top-level Duration/Epoch imports.
